+++ CHANGELOG.md
@@ -6,8 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added 
-- Added send_final_comment to the SendMethod class. This was present in the XML
-  but got lost when we switch to JSON.
+- Added `final_comment` to the SendMethod class. This was present in the XML
+  but got lost when we switched to JSON.
 
 ## [2.0.2] - 2017-06-05
 ### Added